requirements.txt
langchain-openai>=0.0.9
langchain-core>=0.1.32
langchain-community>=0.0.30
langchainhub>=0.1.14
langchain-groq>=0.0.5
langchain-objectbox>=0.0.3

streamlit>=1.30.0
langserve>=0.0.45
fastapi>=0.110.0
uvicorn>=0.29.0
sse-starlette>=1.6.5

python-dotenv>=1.0.1
bs4>=0.0.2
beautifulsoup4>=4.12.3
pypdf>=3.17.1
PyPDF2>=3.0.1

chromadb>=0.4.24
faiss-cpu>=1.8.0
sentence-transformers>=2.6.1
groq>=0.5.1
cassio>=0.1.0

wikipedia>=1.4.0
arxiv>=2.1.0
